requirements.txt
asgiref==3.10.0
Django==5.2.8
djangorestframework==3.16.1
dotenv==0.9.9
gunicorn==23.0.0
mysqlclient==2.2.7
packaging==25.0
python-dotenv==1.2.1
sqlparse==0.5.3
tzdata==2025.2